+++ packages/plugin-giphy/web/src/toolbar/giphyApiInputModal.jsx
@@ -1,10 +1,10 @@
 import PropTypes from 'prop-types';
 import React, { Component } from 'react';
 import { mergeStyles } from 'wix-rich-content-common';
-import { TextSearchInput } from 'wix-rich-content-ui-components';
+import { TextSearchInput, SettingsMobileHeader } from 'wix-rich-content-ui-components';
 import styles from '../../statics/styles/giphy-api-input-modal.scss';
 import GiphySelector from './giphySelector';
-import { CloseIcon } from '../icons';
+import classNames from 'classnames';
 import { debounce } from 'lodash';
 import { GIPHY_TYPE } from '../types';
 
@@ -44,26 +44,25 @@ export default class GiphyApiInputModal extends Component {
     const { styles } = this;
     const { t, isMobile, languageDir } = this.props;
     const searchTag = this.state.searchTag;
-    const mobileNavbar = (
-      <div>
-        <div className={styles.giphy_api_input_modal_navbar}>
-          {t('GiphyUploadModal_mobileNavbar_Title')}
-          <div
-            onClick={this.onCloseRequested}
-            role="button"
-            tabIndex="0"
-            onKeyPress={null}
-            className={styles.giphy_api_input_modal_closeButton}
-          >
-            <CloseIcon />
-          </div>
-        </div>
-      </div>
-    );
+
     return (
-      <div dir={languageDir}>
-        {isMobile ? <div>{mobileNavbar}</div> : null}
-        <div className={styles.giphy_api_input_modal_container} data-hook="giphyUploadModal">
+      <div dir={languageDir} className={isMobile ? styles.gif_modal_mobile_container : null}>
+        {isMobile && (
+          <SettingsMobileHeader
+            theme={this.props.theme}
+            onCancel={this.onCloseRequested}
+            t={t}
+            title={t('GIFPlugin_Mobile_Header')}
+            useNewSettingsUi
+            showSaveBtn={false}
+          />
+        )}
+        <div
+          className={classNames(styles.giphy_api_input_modal_container, {
+            [styles.mobile]: isMobile,
+          })}
+          data-hook="giphyUploadModal"
+        >
           <TextSearchInput
             inputRef={ref => (this.input = ref)}
             onClose={this.onCloseRequested}