+++ packages/ricos-content/web/src/converters/plainText/toPlainText/toPlainText.ts
@@ -9,6 +9,7 @@ import {
   parseTextNodes,
   parseAppEmbed,
   parseVideo,
+  parseAudio,
   parseEmbed,
   parseCollapsible,
 } from './convertNodes';
@@ -16,6 +17,7 @@ import {
 interface PlainTextOptions {
   urlShortener?: (url: string) => Promise<string>;
   getVideoUrl?: (fileId: string) => Promise<string>;
+  getAudioUrl?: (fileId: string) => Promise<string>;
   delimiter?: string;
 }
 
@@ -25,7 +27,7 @@ export const toPlainText = async (
 ): Promise<string> => {
   const ricosContent = RichContent.fromJSON(content);
   const { nodes } = ricosContent;
-  const { urlShortener, getVideoUrl, delimiter = '\n' } = options;
+  const { urlShortener, getVideoUrl, getAudioUrl, delimiter = '\n' } = options;
   let plainText = '';
 
   const parseNodes = async (index = 0) => {
@@ -52,6 +54,9 @@ export const toPlainText = async (
           break;
         case Node_Type.VIDEO:
           plainText += await parseVideo(node, delimiter, getVideoUrl);
+          break;
+        case Node_Type.AUDIO:
+          plainText += await parseAudio(node, delimiter, getAudioUrl);
           break;
         case Node_Type.GIF:
           plainText += parseGiphy(node);