+++ packages/ricos-content/web/src/types/node-refined-types.ts
@@ -20,6 +20,7 @@ import type {
   TableData,
   TextData,
   VideoData,
+  AudioData,
 } from 'ricos-schema';
 
 type Identified = {
@@ -91,7 +92,8 @@ export type RichContentNode =
   | ImageNode
   | LinkPreviewNode
   | MapNode
-  | VideoNode;
+  | VideoNode
+  | AudioNode;
 
 // RichContent Containers
 type RichContainer = {
@@ -212,4 +214,10 @@ export type VideoNode = Identified &
   Leaf & {
     type: Node_Type.VIDEO;
     videoData: VideoData;
+  };
+
+export type AudioNode = Identified &
+  Leaf & {
+    type: Node_Type.AUDIO;
+    AudioData: AudioData;
   };