+++ CHANGELOG.md
@@ -17,6 +17,12 @@
     Click to see more.
   </summary>
 
+
+</details>
+<hr/>
+
+## 8.69.26 (Feb 13, 2022)
+
 ### :house: Internal
 - `schema/audio`
   - [#3461](https://github.com/wix/ricos/pull/3461) add audio schema
@@ -26,8 +32,6 @@
 - `theme`
   - [#3494](https://github.com/wix/ricos/pull/3494) Support Font Family Inheritence
 
-</details>
-<hr/>
 
 ## 8.69.25 (Feb 13, 2022)